code refactor and added postman collection

+++ DynamicApplicationCP/DynamicApplicationCP/Services/QuestionService.cs
@@ -5,51 +5,82 @@ namespace DynamicApplicationCP.Services
 {
     public class QuestionService : IQuestionService
     {
-        private IConfiguration _configuration;
-        private CosmosDBService _cosmosDBService;
-        private string _cosmosDbName;
-        private string _cosmosDbContainerName;
+        private readonly IConfiguration _configuration;
+        private readonly CosmosDBService _cosmosDBService;
+        private readonly string _cosmosDbName;
+        private readonly string _cosmosDbContainerName;
+
         public QuestionService(IConfiguration configuration, CosmosDBService cosmosDBService)
         {
-            _configuration = configuration;
-            _cosmosDBService = cosmosDBService;
-            _cosmosDbName = _configuration["CosmosDB:DatabaseName"];
-            _cosmosDbContainerName = _configuration["CosmosDB:QuestionContainerName"];
+            _configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
+            _cosmosDBService = cosmosDBService ?? throw new ArgumentNullException(nameof(cosmosDBService));
+
+            _cosmosDbName = _configuration["CosmosDB:DatabaseName"]
+                ?? throw new ArgumentNullException("CosmosDB:DatabaseName configuration is missing in appsettings.json");
+
+            _cosmosDbContainerName = _configuration["CosmosDB:QuestionContainerName"]
+                ?? throw new ArgumentNullException("CosmosDB:QuestionContainerName configuration is missing in appsettings.json");
         }
 
         public async Task AddQuestionAsync(QuestionModel questionModel)
-        {            
+        {
+            if (questionModel == null)
+            {
+                throw new ArgumentNullException(nameof(questionModel));
+            }
+
+            questionModel.QuestionId = Guid.NewGuid().ToString();
             await _cosmosDBService.CreateOrUpdateItemAsync(questionModel, questionModel.QuestionId, _cosmosDbName, _cosmosDbContainerName);
         }
 
         public async Task AddMultipleQuestionsAsync(List<QuestionModel> lstQuestionModel)
         {
-            List<Task> concurrentTask = new List<Task>();
-
-            foreach (var item in lstQuestionModel)
+            if (lstQuestionModel == null)
             {
-                item.QuestionId = Guid.NewGuid().ToString();
-                concurrentTask.Add(_cosmosDBService.CreateOrUpdateItemAsync(item, item.QuestionId, _cosmosDbName, _cosmosDbContainerName));
+                throw new ArgumentNullException(nameof(lstQuestionModel));
             }
 
-            await Task.WhenAll(concurrentTask);
+            List<Task> concurrentTasks = new List<Task>();
+
+            foreach (var question in lstQuestionModel)
+            {
+                question.QuestionId = Guid.NewGuid().ToString();
+                concurrentTasks.Add(_cosmosDBService.CreateOrUpdateItemAsync(question, question.QuestionId, _cosmosDbName, _cosmosDbContainerName));
+            }
+
+            await Task.WhenAll(concurrentTasks);
         }
 
         public async Task<string> DeleteQuestionByIdAsync(string questionId)
         {
+            if (string.IsNullOrEmpty(questionId))
+            {
+                throw new ArgumentException("Question ID cannot be null or empty", nameof(questionId));
+            }
+
             return await _cosmosDBService.DeleteDocumentAsync(questionId, _cosmosDbName, _cosmosDbContainerName);
         }
 
         public async Task<List<QuestionModel>> GetQuestionsByProgramIdAsync(string programId)
         {
-            string query = $"SELECT * FROM c where c.programId = '{programId}'";
+            if (string.IsNullOrEmpty(programId))
+            {
+                throw new ArgumentException("Program ID cannot be null or empty", nameof(programId));
+            }
 
+            string query = $"SELECT * FROM c WHERE c.programId = '{programId}'";
             List<QuestionModel> lstQuestionModel = await _cosmosDBService.GetDocumentsAsync<QuestionModel>(query, _cosmosDbName, _cosmosDbContainerName);
 
             return lstQuestionModel;
         }
+
         public async Task UpdateQuestionByIdAync(QuestionModel question)
         {
+            if (question == null)
+            {
+                throw new ArgumentNullException(nameof(question));
+            }
+
             await _cosmosDBService.CreateOrUpdateItemAsync(question, question.QuestionId, _cosmosDbName, _cosmosDbContainerName);
         }
     }
